+++ docs/.vitepress/config.ts
@@ -1,15 +1,14 @@
 import { defineConfig } from "vitepress";
 import { version } from "../../packages/wujie-core/package.json";
-import path from "path";
 const ogDescription = "极致的微前端框架";
 const ogImage = "https://wujie-micro.github.io/doc/wujie.png";
 const ogTitle = "无界";
 const ogUrl = "https://wujie-micro.github.io/doc/";
-const base = process.env.NODE_ENV === "production" ? "/doc/" : "";
+const base = "/doc/";
 
 export default defineConfig({
-  title: "无界",
-  description: "极致的微前端框架",
+  title: ogTitle,
+  description: ogDescription,
   base,
   head: [
     ["link", { rel: "icon", href: `${base}/favicon.ico` }],
@@ -31,14 +30,11 @@ export default defineConfig({
     },
     lastUpdatedText: "最近更新时间",
     socialLinks: [{ icon: "github", link: "https://github.com/Tencent/wujie" }],
-    // algolia: {
-    //   appId: "",
-    //   apiKey: "",
-    //   indexName: "wujie",
-    //   searchParameters: {
-    //     facetFilters: ["tags:en"],
-    //   },
-    // },
+    algolia: {
+      appId: "QMIAJMDLL1",
+      apiKey: "4eaef57a0122ce0454d3aed08f822178",
+      indexName: "wujie-microio",
+    },
 
     footer: {
       message: "Released the MIT License.",
@@ -82,7 +78,7 @@ export default defineConfig({
       "/guide/": [
         {
           text: "入门",
-          collapsible: true,
+          collapsed: true,
           items: [
             {
               text: "介绍",
@@ -92,15 +88,15 @@ export default defineConfig({
               text: "快速上手",
               link: "/guide/start",
             },
-            // {
-            //   text: "定制Demo",
-            //   link: "/guide/demo",
-            // },
+            {
+              text: "创建项目",
+              link: "/guide/install"
+            },
           ],
         },
         {
           text: "指南",
-          collapsible: true,
+          collapsed: false,
           items: [
             {
               text: "预加载",
@@ -150,7 +146,7 @@ export default defineConfig({
         },
         {
           text: "项目实战",
-          collapsible: true,
+          collapsed: true,
           items: [
             {
               text: "vue主应用",
@@ -162,11 +158,21 @@ export default defineConfig({
             },
           ],
         },
+        {
+          text: "插件推荐",
+          collapsed: true,
+          items: [
+            {
+              text: "wujie-polyfill",
+              link: "https://github.com/jardenliu/wujie-polyfill",
+            },
+          ],
+        },
       ],
       "/api/": [
         {
           text: "主应用",
-          collapsible: true,
+          collapsed: true,
           items: [
             {
               text: "bus",
@@ -192,7 +198,7 @@ export default defineConfig({
         },
         {
           text: "子应用",
-          collapsible: true,
+          collapsed: true,
           items: [
             {
               text: "wujie",
@@ -205,7 +211,7 @@ export default defineConfig({
       "/pack/": [
         {
           text: "框架封装",
-          collapsible: true,
+          collapsed: true,
           items: [
             {
               text: "Vue组件封装",