+++ docs/wujie/components/wujie-connect.vue
@@ -17,7 +17,7 @@ const props = defineProps<{
 watch(
   () => props.baseUrl,
   (newValue) => {
-    url.value = newValue;
+    url.value = newValue!;
   }
 );
 </script>
@@ -41,7 +41,7 @@ watch(
         </form>
       </div>
       <p class="help">
-        您可以输入一个允许跨域访问的<span class="link">https</span>协议网站来在线体验<span class="link">无界</span>
+        您可以输入一个允许跨域访问的 <span class="link">https</span> 协议网站来在线体验 <span class="link">无界</span>
       </p>
     </div>
   </section>