+++ packages/wujie-vue3/index.d.ts
@@ -1,12 +1,11 @@
 import { bus, preloadApp, destroyApp, setupApp } from "wujie";
-import { App } from "vue";
+import type { DefineComponent, Plugin } from 'vue';
 
-declare const WujieVue: {
+declare const WujieVue: DefineComponent & Plugin & {
   bus: typeof bus;
   setupApp: typeof setupApp;
   preloadApp: typeof preloadApp;
   destroyApp: typeof destroyApp;
-  install: (app: App) => any;
 };
 
 export default WujieVue;