+++ docs/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.16](https://github.com/Tencent/wujie/compare/v1.0.15...v1.0.16) (2023-05-17)
+
+**Note:** Version bump only for package wujie-docs
+
+## [1.0.15](https://github.com/Tencent/wujie/compare/v1.0.14...v1.0.15) (2023-05-04)
+
+### Features
+
+* **docs:** add create-wujie md ([#537](https://github.com/Tencent/wujie/issues/537)) ([99c191a](https://github.com/Tencent/wujie/commit/99c191a9f0326b6f9ca7b106ed37ce264e892d2c))
+
+## [1.0.14](https://github.com/Tencent/wujie/compare/v1.0.13...v1.0.14) (2023-03-31)
+
+**Note:** Version bump only for package wujie-docs
+
+## [1.0.13](https://github.com/Tencent/wujie/compare/v1.0.12...v1.0.13) (2023-03-29)
+
+**Note:** Version bump only for package wujie-docs
+
+## [1.0.12](https://github.com/Tencent/wujie/compare/v1.0.11...v1.0.12) (2023-03-28)
+
+**Note:** Version bump only for package wujie-docs
+
+## [1.0.11](https://github.com/Tencent/wujie/compare/v1.0.10...v1.0.11) (2023-03-24)
+
+**Note:** Version bump only for package wujie-docs
+
+## [1.0.10](https://github.com/Tencent/wujie/compare/v1.0.9...v1.0.10) (2023-03-23)
+
+**Note:** Version bump only for package wujie-docs
+
+## [1.0.9](https://github.com/Tencent/wujie/compare/v1.0.8...v1.0.9) (2023-03-23)
+
+### Bug Fixes
+
+* 修复子应用html默认属性 ([#462](https://github.com/Tencent/wujie/issues/462)) ([08e6fd2](https://github.com/Tencent/wujie/commit/08e6fd27b8f5facf4e11bdfdc3426706dca646f9))
+
+### Features
+
+* **doc:** add algolia search ([#447](https://github.com/Tencent/wujie/issues/447)) ([c8c022a](https://github.com/Tencent/wujie/commit/c8c022ac4b61377d7711d33fb19eef2fe5760dde))
+
+## [1.0.8](https://github.com/Tencent/wujie/compare/v1.0.6...v1.0.8) (2023-03-14)
+
+**Note:** Version bump only for package wujie-docs
+
+## [1.0.7](https://github.com/Tencent/wujie/compare/v1.0.6...v1.0.7) (2023-03-14)
+
+**Note:** Version bump only for package wujie-docs
+
 ## [1.0.6](https://github.com/Tencent/wujie/compare/v1.0.5...v1.0.6) (2023-02-08)
 
 **Note:** Version bump only for package wujie-docs